Delegate the handling of default routes to another module
